# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Geld.Network Server.",
        "title": "Geld.Network API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/masterdata/remove/{_id}": {
            "delete": {
                "description": "Delete an existing Master Data Container By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master data container"
                ],
                "summary": "Delete Master Data Container By ID",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/masterdata/save": {
            "post": {
                "description": "This creates a container to whole master data records per type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master data container"
                ],
                "summary": "Save Master Data Container submitted by Geld Configurations",
                "parameters": [
                    {
                        "description": "Master Data Container Details",
                        "name": "masterDataContainerBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MasterData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/masterdata/{_id}": {
            "get": {
                "description": "Get an existing Master Data Container ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master data container"
                ],
                "summary": "Get Master Data by ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/masterrecord/{dataid}": {
            "get": {
                "description": "Retrieves paginated data collection associated with a specific dataID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Get Paginated data collection",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "DataID",
                        "name": "dataid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default from env: PAGINATION_DEFUALT_LIMIT)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Requested page (default from env: PAGINATION_DEFAULT_PAGE)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sort order (-1: Desc, 1: Asc, default from env: PAGINATION_DEFAULT_SORT)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/record/remove/{_id}": {
            "delete": {
                "description": "Delete an existing Master Data Record By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Delete Master Data Records By ID",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/record/save": {
            "post": {
                "description": "This creates a collections for Master Data Containers based on container type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Save Data Collection for Master Data Container",
                "parameters": [
                    {
                        "description": "Data Collection Details",
                        "name": "dataCollectionBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DataCollection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/record/{_id}": {
            "get": {
                "description": "Get an existing Record Data By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Get Record Data By ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/records/{dataid}": {
            "get": {
                "description": "Get an existing Record Data By Master Data ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Get Record Data By Master Data ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DataID",
                        "name": "dataid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/updatemasterdata": {
            "put": {
                "description": "Update an exisiting Master Data Container",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master data container"
                ],
                "summary": "Update Master Data Container",
                "parameters": [
                    {
                        "description": "Master Data Container Details",
                        "name": "masterDataContainerBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestDtos.UpdateMasterData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/updaterecords": {
            "put": {
                "description": "Update an exisiting Master Data Container",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data collection"
                ],
                "summary": "Update Master Data Container",
                "parameters": [
                    {
                        "description": "Data Collection Details",
                        "name": "dataCollectionBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestDtos.UpdateDataCollection"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/usermasterdata/{userid}": {
            "get": {
                "description": "Retrieves paginated master data associated with a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "master data container"
                ],
                "summary": "Get Paginated master data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default from env: PAGINATION_DEFUALT_LIMIT)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Requested page (default from env: PAGINATION_DEFAULT_PAGE)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sort order (-1: Desc, 1: Asc, default from env: PAGINATION_DEFAULT_SORT)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responseDtos.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.DataCollection": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "collectionname": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "dataid": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "purpose": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
        "model.MasterData": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "dataname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "requestDtos.UpdateMasterData": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "dataname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "userid": {
                    "type": "string"
                }
            }
        },
...
        "responseDtos.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "responseDtos.ResultResponse": {
            "type": "object",
            "properties": {
                "Response": {},
                "Status": {
                    "type": "integer"
                }
            }
        }
    }
}